Update Proto_1.py
I have inserted everything into a canvas to allow for scrolling capabilities, I have added the scrollbars and created a simple placeholder version of the quiz system after much difficulty due to forgetting to enter the master parameter. Next is to either continue work on this first prototype version or work on the next prototype. If continuing on this version than the last thing to do is to add randomness to the quiz and create a way for it to move onto the next question.